Item Settiing
- 무기 관련 에셋 추가
- 아이템 스프라이트 가공 및 기본세팅
- 회피율 적용

+++ Assets/Scripts/Enemy/EnemyAI.cs
@@ -26,7 +26,7 @@ public class EnemyAI : MonoBehaviour
     [SerializeField] private GameObject onTriggerCheckObj;
     [SerializeField] private TriggerCheck triggerCheck;
 
-    public int aggroMeter1, aggroMeter2;  // 어그로미터기
+    public float aggroMeter1, aggroMeter2;  // 어그로미터기
 
     // Start is called before the first frame update
     void Start()
